+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Changed
 - Remove node operator
+- Skip removing genesis tips
 
 ## [v2.19.0] 2020-11-16
 ### Added